Move query for finding due reminders to the model
This commit refactors how review reminders that are due a reminder email
notification get found by the 'send review reminders' Rake task.

I've moved the custom query out of the Rake task and into a scope on the
ReviewReminder model. This feels like a more sensible long-term location
for the query, as it keeps that business logic together with the model
itself, and reduces the amount of 'knowledge' held in the Rake task.

I've introduced various traits on the ReviewReminder factory so that
example objects can easily be created in a known state within tests.

One other subtle change here is that I've updated the query to correctly
handle the new DATE type of the review_at column. This required changing
the query to find records where review_at is less than *or equal to*
today. Previously the query was only looking for items due for review
*less than today* (i.e. yesterday and prior). The original query only
worked with the previous DATETIME fields when London was in the BST
timezone. See the previous commit for more details.

The other subtle change is that I've removed the call to `.to_s` on the
Document ID which was converting the integer into a string. Sidekiq
workers can accept integer parameter values just fine, so this wasn't
necessary. (However, they only accept string parameter keys. I'm
assuming that `.to_s` was originally implemented as a misunderstanding
of that 'keys must now be strings' requirement.)