Add PUT route for updating document collection email override
If the override is succesfully set, redirect to the document collection
edit page. In all other cases, reload to the email_override_edit page and
show a relevant message.

It should only be possible to set the override on a subset of draft document
collections - this restriction will be added in subsequent commit.