Move Timecop.freeze to where it's needed
It wasn't immediately obvious to me why this test needed to use
Timecop.freeze. After giving it some thought, I think it's needed so
that time doesn't progress between running the worker and asserting that
it correctly set the current time.

I've added a comment to explain that, and moved the Timecop.freeze to
specifically where it's needed.