Make test more realistic
I've made this test more realistic to expose a bug that has been
observed in the real world.

In the real world, the review_at date will be in the past when the
worker tries to set the reminder_sent_at timestamp.

But the review_date_cannot_be_in_the_past validation rule has made it
impossible for the worker to set the reminder_sent_at timestamp. Instead
it's been put in the retry queue (due to an unhandled validation
exception) resulting in mail notifications being repeatedly sent to
users in quick succession.