Bypass validation when setting reminder_sent_at
We need to bypass validation rules when updating the reminder_sent_at
field. Otherwise the record is condidered invalid due to the review_at
timestamp being in the past.

Needing to bypass this validation rule feels like a bit of a code smell.
Ideally the model wouldn't raise validation errors in this scenario.

I'm applying this as a 'quick fix' because we need to ensure the job
runs smoothly before the daily cron next runs. 😬 But ideally I'd like
to follow up with some more changes to refactor the Review Reminder
model slightly.