Migrate DATETIME to DATE for document review dates
This migration converts the review_at DATETIME field to just DATE.
Users can only choose a date when populating this field, so DATE is a
more appropriate column type to use. Users cannot specify a time, so all
times have defaulted to "00:00:00" on the specified date.

However, since Ruby's DateTime objects are localised to the current
London timezone, it means dates that were entered during BST will have
been stored incorrectly in the database. This is because Rails stores
DateTime objects as UTC times in the database.

For example, during BST a user input of "2023-08-01" would have become a
DateTime object with the value "2023-08-01 00:00:00 +0100".  The "+0100"
represents the fact that the current timezone is BST (which is UTC + 1
hour). When written to the database, this would have been stored as UTC
time "2023-07-31 23:00:00".

This has resulted in review dates being stored incorrectly as the day
before the user actually specified. 😬

To resolve this, I'm converting the timezone of these DATETIME fields
prior to converting them to DATE fields. I'll add +1 hour so they revert
back to the date users originally specified.

For completeness, the rollback will convert DATE values back to their
previous (albeit inaccurate) DATETIME values.