Don't send reminder if it's not due yet
This adds a method ReviewReminder#review_due? and uses it in the
ReviewReminderNotifierWorker to ensure that reminder emails aren't sent
if they're not due.

This tightens up the existing guard clause which already protected
against send reminders which have already been sent. Now it will also
skip reminders if they're queued too early.

I've done this primarily to avoid the likelihood of us creating tests
that inadvertently behave differently than in 'real life' because they
rely on future review_at dates. This actually happened and was resolved
in PR #8047 - but this is a follow-up refactor which should make the
worker behave in a 'obvious' way.