Introduce ReviewReminder#reminder_sent! method
As a more user-friendly way to mark a review reminder as having been
sent.

External classes now no longer need to know about the `reminder_sent_at`
timestamp directly to find items that are due a reminder, nor to record
it as sent.

I've also made a subtle change to the validation rule that ensures the
review_at date is not in the past. I've done this so it's possible to
update reminder_sent_at date without triggering a validation error - due
to the review_at date technically being in the past by then.

Previously I'd used ActiveRecord::Persistence.update_columns to update
that column without validations. However this feels like slightly
unexpected behaviour because it also means callbacks are skipped. It
feels cleaner to be able to update this field without having to bypass
validations.