Refactor `PublishingApiPresenters`
We were using a very lengthy `case-when` block to determine which
Publishing API presenter to use for a given model.

Now all models are published to Publishing API, we don't need this
complexity and can move the presenter onto the model itself.

This should make the code easier to work with, as
`PublishingApiPresenters` now won't need to be amended when a new model
is added or an existing model renamed.