Update `PublishingApiDocumentRepublishingWorkerTest`
This test was only working because the logic in
`PublishingApiPresenters` meant we raised an error there as the object
did not match any of the types in the `case-when` block.

Therefore updating the test to use an actual document and edition so we
have all the methods required to attempt to republish the document.