Add an index to optimise database queries
I've added an index to the `review_reminders` table to optimise for the
common read pattern of finding reminders that are due to be sent.

I've used a ["composite index"][1] (also known as a "multiple-column index")
to index both the `review_at` and `reminder_sent_at` columns together.
This enables finding by `review_at` alone (since it's the first column
in the index), or by `review_at` and `reminder_sent_at` combined.

In particular, this optimises the database for handling queries produced
by the `ReviewReminder.reminder_due` scope, which has conditions on both
these fields.

We can tell this index is working as expected by looking at the "EXPLAIN
statement" for the query. In a Rails console, we can run:

```
ReviewReminder.reminder_due.explain
```

This outputs quite a lot of information, and it can take a while to get
your bearings and take in what it's telling you. The [MySQL docs][2] go
into depth about what all the output means.

However for our purposes, we want to look at the value of the "key"
column. This tells us which index MySQL would use when performing this
query.

Before applying this migration, the "key" column contained NULL. This
means MySQL couldn't identify a key that'd be useful in optimising this
query. It'd therefore have to fallback to a complete scan of the table
to find matching records (which can be very slow on big tables).

After applying this migration, the "key" column contained the new index
`index_review_reminders_on_review_at_and_reminder_sent_at`. This means
MySQL will use the index when performing that query – so we've
successfully optimised for this read pattern.

[1]: https://dev.mysql.com/doc/refman/8.1/en/multiple-column-indexes.html
[2]: https://dev.mysql.com/doc/refman/8.0/en/explain-output.html#explain-output-columns