Fix `StackLevelTooDeep` error when adding images to news articles
Following the deployment of #8100 we have seen `StackLevelTooDeep` errors where
the `Edition::LeadImage` module included in the `NewsArticle` model is
recursively calling the `PublishingApi::NewsArticlePresenter`.

This is happening because `image_url` which is defined in the
`Edition::LeadImage` is overridden by `NewsArticle`.

To fix, we can remove the override, it's only called in one place which can be
replaced with a more direct call instead.