Republish Worldwide Organisation on role update
The contents of the role (e.g. the name of the role) are included in the
Worldwide Organisation pages.

Therefore we need to republish them when the content updates.

We do not need to republish when a role is deleted, as this will be
captured by the `after_destroy` on the `WorldwideOfficeRole` model.