Save asset for promotional feature - nested items images
Implementing saving assets for promotional feature with nested items in order to
move away from using legacy_url_path for talking to asset manager. The asset model
will provide an asset_manager_id to facilitate this instead.

We are setting the user permission in the feature controller and expecting the
model in asset manger storage to correctly identify it.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk