Remove `/government` as a route that can be published
This route was previously used as a "catch all" for any content that was
to be rendered by Whitehall that didn't have a content item.

As all content now has a content item and Whitehall renders no content,
we have published this route (as a redirect) in Short URL Manager.

Therefore we no longer need to publish this route from Whitehall.

The other two remaining Special Routes should also be moved to Special
Routes Publisher, as they do not require any data from Whitehall's
database, but that is outside the scope of this work.