Add current release date FK association to statistics announcement table
We've recently run into an issue on the statistics announcements
index page where ordering statistics announcments based on the release
date isn't working correctly.

It's difficult to isolate the current_release_date for a statistics
announcements due to it being a non-performant has_one association that
orders release dates by DESC then selects the first one.

We've decided to treat this in the same way we did editions where
we will persist the current_release_date's id as a column that lives on
the statistics announcement. We can then join the announcements via the
association ensuring that only the current_release_date is used when
ordering by the release date.