Remove orphaned search index records belonging to editions which are no longer searchable
There are a number of documents appearing in search results which are written in a non-English language (we're aware of many Welsh language documents but there could be others). This migration finds all publicly visible documents which should not be searchable (we expect this to be largely non-English language documents) and attempts to ensure they are removed from the search index. Testing in the integration environment suggests that this does not remove all of the orphans but it has tidied up at least some of them.