Add correct data migration for removing orphaned search index records
This is a correct implementation of 402f27c32721100a1c504416c8400c1db21c70ba. There are a number of documents appearing in search results which are written in a non-English language (we're aware of many Welsh language documents but there could be others). This migration finds all publicly visible documents which should not be searchable (we expect this to be largely non-English language documents) and attempts to ensure they are removed from the search index. Testing in the integration environment suggests that this does not remove all of the orphans but it has tidied up at least some of them.