Add migration to backfill current_release_date_id
This backfills the new column. Unfortunately, we need to use update_column
here to skip validations as quite a few announcments are invalid.

It may well be better anyway though as if we used update, we would
essentially be bulk updating all announcments and sending them off to
the Publishing API.