Update current_release_date assoication to use FK
This updates the statistics_announcements model to use the new backfilled
FK rather than mnaually sorting the statistics_announcement_dates
based on the most recently created.

This will both be more performant and allow us to more easily scope and
order DB queries based on the current_release_date of the
statistics_announcement.

Using this in place of the current_release_date assoication in the
factory requires us to update some tests to use create rather than build
due to various callbacks occuring that require the current_release_date_id
to be persisted and the assoication reloaded.