Ensure document slug can never change once an edition has been published
Without this change, updating the title on the draft edition created after the previous edition was unpublished will cause the document slug to update. This causes confusing behaviour for publishers, who are probably not aware that the document slug has changed when they update the title of the draft. This has lead to them reporting that unpublishing is not working correctly, because they have expected to see the unpublishing notice or redirect at the new slug rather than the old one, especially as they can easily reach the new slug by previewing on the draft stack and then changing the domain to the live stack in their browser address bar. Also, we probably don't want a re-published document to appear at a different slug, as documents are intended to be uniquely identifiable.