Update statistics_announcements to accept nested attrs for statistics_announcements_dates
We need to update the statistics_announcements model to accept attrs
for a statistics_announcement_date rather than the current_release_date
now that the current_release_date_id assoiciation uses a foreign key.

You can no longer set this manually via the factory as it requires the
object to be persisted and the id set. This means that we need to pass
it through via the statistics_announcement_dates assoication for it
to be persisted to the DB.