Render the About us page using edition rendering
The code for rendering an edition does all the necessary processing of
the attachments. In order to use the edition rendering, we need to pass
the About us object to the rendering method, instead of its body.

I've added a test because this functionality is quite complex, relying
as it does on special treatment of a worldwide corporate info page.