Retain a users org when resetting search fields on document search page
At the moment, when you're searching for a document and click the
"Reset all fields" it doesn't retain the users organisation and
shows documents from all organistions.

As the filter_options partial is reused across multiple views, some of
which don't have the option to filter by organisation, we need to ensure
that only the document search page is affected.

This adds in a helper method which handles this.