Move flash messages to view component and add html_safe option
As part of the work to allow flash messages to have a html_safe option
we have moved the flash message in the design system layout to be put
together by a view component.

We then added an option to flash messages to allow them to be marked as
html_safe. If a flash message is marked as safe it will correctly render
any custom html.

Previously we always marked flash[:alert] messages as safe and not
flash[:notice] messages. Since this was no documented we could have
added a flash[:alert] with user input by accident which is a security
risk.

This work should make the application safer by making us make a
deliberate choice any time we render a flash message.