Ensure correct ordering of for previous_display_date
This fixes an issue where previous_display_date was causing
statistics announcments dates to return in an incorrect order due
to one being built and the other being persisted to the DB.

Due to this the persisted one was returning first, despite being
second when reloaded.

While we could reload the statistics_announcement in the test before
passing it into the presenter, it feels like an issue could well
run into down the line and cause more headaches to i think it makes
more sense to just perisist the it in the factory.

It's not an overly used transient so it won't expect the performance
of the test suite.