Add confirm_destroy endpoint in the GOV.UK Design System
I've chosen to just redirect to the index page if the group cannot
be deleted because there's only 1 group on the document collection.

This page would only be accessible if the user manually visits the
page as the delete link only shows if 2 groups are present so it's
unlikely to occur anyway.