Add tabbed nav helper method for document_collection_groups
This adds a helper method which handles rendering secondary navigation
sub navs for document_collection_groups.

At the moment, there's no href for the documents sub nav as a the
document collection group membership index endpoint has not been
defined in the routes file. It'll be added in a follow up PR.