Fix failing test by making it less brittle
The checkbox label for access limiting has changed, which resulted in
this integration test breaking.

To resolve it, I've made the test more flexible so it'll check the label
containing "Limit access", since I think this is still specific enough
to avoid being confused with any other checkbox on the page.