Add the DocumentCollectionGroupMemberships#index endpoint
This adds the index endpoint for document collection group
memberships. This page links off to the add document flow (yet to be
built) and either warns the user there are no documents in the group,
or displays a table of documents. Each of these links to the public URL
and the confirm delete page (also still needs to be added).