Conditionally render a reminder to add tags.
Now that we have removed tags from the editor workflow (see previous
commit). We should ensure the publishers know to add tags/taxons to
their documents.

We need to display the link to updating tags in the flash message when
there are no tags and the user have clicked on the "Save and go to
document summary" button from the document creation/editing page.