Added styling for Document Collection Group index page
Updated GovUK tables' cells on the document collection page so that they are now word wrapping in case we have a document with a really long title, so that it doesn't break the page layout. Added a fixed width to the right hand column so that View and Delete links don't stack when there is a long title.

Co-Authored-By: David Gisbey <david.gisbey@digital.cabinet-office.gov.uk>