Add option to filter documents by overdue review reminders
This adds in an additional filter to the documents search page which
allows users to be able to filter by review reminders that are overdue.

Unforunately i've had to manually adds the joins to the documents and
review reminder tables due to some custom sql written in the
:with_title_containing scope

```
in_default_locale
  .includes(:document)
  .where("edition_translations.title LIKE :like_clause OR documents.slug = :slug", like_clause:, slug: keywords)
  .references(:document)
```

This expects the documents to be called documents, but when you use
active record to join a has one relationship it aliases the table to
singular.

So in this case, when you join the document it aliases the documents
table to 'document' which casuses the above where clause to blow up
as it can't find the documents table. Interestingly the same behaviour
doesn't seem to occur for left outer join.