Block publishing if all asset variants have not finished uploading
Current behaviour is that the user cannot publish a document unless the original variant (e.g. only the original file if you have a pdf, we don’t care about thumbnail) has been uploaded to asset manager.  This checks the uploaded_to_asset_manager_at timestamp on the AttachmentData.
We extend the checking to all variants so that it does not only cares about original file now.

Co-authored-by: Jessy Yeh <jessy.yeh@digital.cabinet-office.gov.uk>