Save use_non_legacy_endpoints user permission as a field in the database
The original approach was to keep this field on the AttachmentData as
an in memory field. Nonetheless, we ended up requiring it in places
where it doesn't travel from the controller where we had originallly
set it on an AttachmentData instance. Now saving it in the database
so it is available.

This will be removed once the feature flag is not needed anymore.