Add Rake tasks to run and verify the migration
I've added a Rake task to perform the migration of Consultations to
Calls for Evidence, and another which will verify that the expected
redirects are in place and content is being served from the new URL.

Environment variables can be set for these Rake tasks:

MIGRATION_CSV_PATH - defaults to the CSV in data_migration
MIGRATION_USER_ID - the User ID to attribute the migration to,
                    defaults to the "GDS Inside Government Team" user
GOVUK_WEBSITE_ROOT - already set in EKS hosted environments

1. Run the migration with:

MIGRATION_USER_ID=1234 bundle exec rake migrate_calls_for_evidence:migrate

It will output a list of failed migrations accompanied by the error
message of the exception that was thrown.

Output is formatted as CSV so you should be able to copy & paste the
results from your terminal and analyse them in a spreadsheet app.

2. Verify the migration with:

bundle exec rake migrate_calls_for_evidence:verify

This will do a quick sanity check so we can be confident the migration
worked successfully. It performs HTTP requests for the original
Consultation URLs, and expects to be redirected to a Call for Evidence
URL with the same slug.

Any unexpected HTTP responses will be output - again in CSV format, to
make it easier to analyse the results in a spreadsheet app.