Fix filename and url to work with asset IDs
Now sending non-legacy urls in the form "/media/:id/:filename to the publishing api, for
attachments created with the use_non_legacy_endpoints permission.

There are draft updates being sent to the publishing api, potentially  before the assets
are available. In order to adress this, we are filtering out unwanted attachments from
the load - from both the govspeak and the attachments builder payload.

The new media url is used for rendering views within Whitehall as well, posing the same issue
of assets not being immediately available. Currently, we are returning a nil url if assets
are not available.

The path method ends up being called (carrierwave internals) if the url method returns nil.
We have tried to remove the path method from the File class altogether, but a lot of existing
functionality relies on it, primarily around asset deletion for non file attachments. Therefore,
path  will have to stay as long as we are still working on attachments for the other content types.
That means we are returning @legacy_url_path for nil urls. The views will be fixed in a future commit
so that they render user-friendly output, meaning the nil urls should not be visible to any users.

Previously, filenams was obtained from the url, which relied on @legacy_url_path. The identifier
path for files is always available and can be used to extract filename. Filename is now directly
extracted from this path, thus decommissioning another use of @legacy_url_path.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk