Add translated_textarea component
At the moment, we have quite a few translation pages that render textareas
and inputs then the english translation below them.

We've decided that we're going to migrate the h2 and english translation
for both to a details component.

While doing this I think it makes sense to add reusable components
which handle spacing etc. so they're all consistent.

This adds a 'translated_textarea' component. A follow up commit will
add a 'translated_input' one.