Changed title lookup scope for editions to work with ActiveRecord associations
Unfortunately when we do associations between edition and document, the singularisation causes mismatch between table names. When we craft queries ActiveRecord aliases the table name from documents to document, which seems like any query with a mixed usage of documents and document looked like it was throwing an error. We fix this by changing the Edition Title Slug query to an active record query rather than direct SQL and it seems to work.

Had to additoinally pull out scope in_default_locale.includes(:document) as a common set of scope so both the where clauses are looking at the same thing. Checked the console, it's still only doing 1 query to the DB.