Trigger edition draft updater when assets are uploaded
This draft updater was originally only called from the AttachmentsController.
Since we are now using asset IDs (rather than legacy url paths) to compose the
attachments' urls, we need to send these urls to the publishing api once
they become available. The place to do so is inside the worker that
saves the assets in the first place.

The regular call to draft updater from within AttachmentController
is still happening, but the load will not include any attachments without
assets. This filtering is done in the PayloadBuilder for attachments at the
presenter level as well as inside the govspeak helpers.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk
Co-authored-by: Tuomas Nylund tuomas.nylund@digital.cabinet-office.gov.uk