Extend publishing validation to not allow publishing if attachments assets are missing
Previous behaviour was to not allow publication if any of the attachments of an attachable
have not finished uploading - strictly speaking, their original asset has not finished uploading.
Any pending variants, e.g., a thumbnail, were ignored in this validation.

Since we are now implementing asset IDs instead of legacy paths, we expect all assets to have
been created and we disallow publishing if any of the asset variants are missing.

Co-authored-by: Syed Ali syed.ali@digital.cabinet-office.gov.uk
Co-authored-by: Tuomas Nylund tuomas.nylund@digital.cabinet-office.gov.uk