Ensure scheduled publication is set to a future date
Now that we're allowing users free rein to enter their dates using text inputs, we need to start validating that they make sense in the context of the application. This only applies to draft publications, as it is perfectly valid for a published edition to have a scheduled publication date in the past.