Fix bug preventing hour and minute date time components being set to 0
Rails expects the date to be provided as a hash. We are converting the date values to integers (we decided this was simpler than using regexes) to ensure that non-numeric values are flagged as invalid. When the value is non-numeric the result will be 0. However, we accidentally applied the check to the time elements of the date hash as well, causing a bug where it was not possible to set the minute or hour values to 0.