Modify scheduled publication date validation error message
The updated message is closer to the guidance provided at https://design-system.service.gov.uk/components/date-input/, but the guidance doesn't cover cases where the user enters, for example, a string instead of a number for one of the date parts. It also assumes that invidividual parts of the date can be validated separately. For this reason I have had to improvise a little, but I believe the new message is closer to the spirit of the guidance and does not feature the unapproved XX style formatting advice.