Ensure ministerial reordering publishes changes to Publishing API
At the moment, when you reorder ministerial roles and organisations,
it calls update_all which skips validations and callbacks.

This means that the after_save which pushes changes to Publishing API
is not called, so the changes are not passed downstream.

We've decided to retain the behaviour where validations are skipped as it
feels like this functionality shouldn't be blocked via an invalid role
or organisation.

However, save!(validate: false) does trigger the callbacks so the roles
organisations will continue to be republished.