Remove temporary whodunnit race condition test
This is just a housekeeping PR to remove a test that was useful in proving we had resolved an issue relating to incorrect whodunnit values caused by sharing state between threads. Since implementing the fix in 1d6df6f628925077e517297426e4952f96b1c30a we haven't seen any reoccurences of the issue, so it's probably safe to delete the test now.