Ensure org is republished after reordering people
At the moment, update_column is used to ensure that the uniq validation
on ordering is not triggered while the updates take place.

However, this also skips any callbacks which handle republishing the
org.

I think it makes sense to retain this behaviour and manually republish
the org, so the org isn't republished for each role that is reordered.