Use govuk publishing date component for review reminder fields
Continuation of work to replace all Whitehall select-based date inputs with the GOVUK publishing date component. Validation is now needed for the date but will follow in a separate commit. The new component uses text inputs, hence the addition of a new feature test helper method for filling in the input. Eventually we will replace the old helper.